Add Major Figures
The Figures are rendered in Python files seperate from the Quarto file since some of XGBoost's classes seem to be incorrectly deleted, resulting in errors. The figures are still completed when errors are present as the errors appear during the deletion of variables (after they have been used), however Quarto does not seem to process the document further when any errors are present. This fix notably also simplifies the reading of the Quarto file.